+++ core/repository.go
@@ -28,3 +28,27 @@ type GitRepositoryConfig struct {
 	// RootDir is the local root path to the Git repository.
 	RootDir string
 }
+
+// GitRepositoryStore is a core service that is responsible for providing services for managing Git repositories.
+//counterfeiter:generate . GitRepositoryStore
+type GitRepositoryStore interface {
+	// FetchGitRepositories will load the managed repositories from a config store and returns an array of GitRepository for each Git repository.
+	// A non-nil empty slice is returned if there are none existing.
+	FetchGitRepositories() ([]GitRepository, error)
+}
+
+// Label is attached to a remote Git repository on a supported Git hosting provider.
+// The implementation may contain additional provider-specific properties.
+//counterfeiter:generate . Label
+type Label interface {
+	// IsInactive returns true if the label is bound for removal from a remote repository.
+	IsInactive() bool
+	// GetName returns the label name.
+	GetName() string
+
+	// Delete removes the label from the remote repository.
+	Delete() (bool, error)
+	// Ensure creates the label in the remote repository if it doesn't exist.
+	// If the label already exists, it will be updated if the properties are different.
+	Ensure() (bool, error)
+}
